style: move plugin_actions.py to the gui folder
